+++ kuksa_databroker/databroker/src/grpc/server.rs
@@ -1,9 +1,22 @@
+/********************************************************************************
+* Copyright (c) 2022 Contributors to the Eclipse Foundation
+*
+* See the NOTICE file(s) distributed with this work for additional
+* information regarding copyright ownership.
+*
+* This program and the accompanying materials are made available under the
+* terms of the Apache License 2.0 which is available at
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* SPDX-License-Identifier: Apache-2.0
+********************************************************************************/
+
 use std::{future::Future, time::Duration};
 
 use tonic::transport::Server;
 use tracing::info;
 
-use databroker_api::{kuksa, sdv};
+use databroker_proto::{kuksa, sdv};
 
 use crate::broker;
